+++ src/libAtomVM/stacktrace.c
@@ -19,9 +19,12 @@
  */
 
 #include "stacktrace.h"
+
+#include "bif.h"
 #include "defaultatoms.h"
 #include "globalcontext.h"
 #include "memory.h"
+#include "nifs.h"
 
 #ifndef AVM_CREATE_STACKTRACES
 
@@ -30,6 +33,19 @@ term stacktrace_create_raw(Context *ctx, Module *mod, int current_offset, term e
     UNUSED(ctx);
     UNUSED(mod);
     UNUSED(current_offset);
+    return exception_class;
+}
+
+term stacktrace_create_raw_mfa(Context *ctx, Module *mod, int current_offset, term exception_class, term module_atom, term function_atom, int arity)
+{
+    UNUSED(ctx);
+    UNUSED(mod);
+    UNUSED(current_offset);
+    UNUSED(exception_class);
+    UNUSED(module_atom);
+    UNUSED(function_atom);
+    UNUSED(arity);
+
     return exception_class;
 }
 
@@ -87,6 +103,11 @@ static bool location_sets_append(GlobalContext *global, Module *mod, const uint8
 }
 
 term stacktrace_create_raw(Context *ctx, Module *mod, int current_offset, term exception_class)
+{
+    return stacktrace_create_raw_mfa(ctx, mod, current_offset, exception_class, UNDEFINED_ATOM, UNDEFINED_ATOM, 0);
+}
+
+term stacktrace_create_raw_mfa(Context *ctx, Module *mod, int current_offset, term exception_class, term module_atom, term function_atom, int arity)
 {
     if (term_is_nonempty_list(ctx->exception_stacktrace)) {
         // there is already a built stacktrace, nothing to do here
@@ -172,19 +193,52 @@ term stacktrace_create_raw(Context *ctx, Module *mod, int current_offset, term e
 
     free(locations);
 
+    // there is an additional x register available as a temporary storage, we'll use it
+    // we'll backup the exception_reason in it
+    int live_x_regs = MINI(arity, MAX_REG) + 1;
+    ctx->x[live_x_regs - 1] = ctx->exception_reason;
+
     // {num_frames, num_aux_terms, filename_lens, num_mods, [{module, offset}, ...]}
-    size_t requested_size = TUPLE_SIZE(6) + num_frames * (2 + TUPLE_SIZE(2));
+    size_t requested_size = TUPLE_SIZE(6) + num_frames * (2 + TUPLE_SIZE(2)) + LIST_SIZE(1, arity) + TUPLE_SIZE(5);
     // NOLINT(term-use-after-gc) exception_class is always an atom
-    if (UNLIKELY(memory_ensure_free_with_roots(ctx, requested_size, 1, &ctx->exception_reason, MEMORY_CAN_SHRINK) != MEMORY_GC_OK)) {
+    if (UNLIKELY(memory_ensure_free_with_roots(ctx, requested_size, live_x_regs, ctx->x, MEMORY_CAN_SHRINK) != MEMORY_GC_OK)) {
         fprintf(stderr, "WARNING: Unable to allocate heap space for raw stacktrace\n");
         return OUT_OF_MEMORY_ATOM;
     }
 
+    ctx->exception_reason = ctx->x[live_x_regs - 1];
+
     term raw_stacktrace = term_nil();
 
-    term frame_info = term_alloc_tuple(2, &ctx->heap);
-    term_put_tuple_element(frame_info, 0, term_from_int(mod->module_index));
-    term_put_tuple_element(frame_info, 1, term_from_int(current_offset));
+    term frame_info;
+
+    // on OTP <= 22 module_atom is set to erlang, when calling a function such as erlang:throw
+    // making this heuristic unreliable since hides the throw caller from the stacktrace
+    // this means that either this heuristic is not 100% correct, or something changed in OTP-23
+    // anyway on OTP >= 23 seem to work as expected.
+    if (module_atom == UNDEFINED_ATOM) {
+        // module_atom has not been provided, let's use mod->module_index
+
+        frame_info = term_alloc_tuple(2, &ctx->heap);
+        term_put_tuple_element(frame_info, 0, term_from_int(mod->module_index));
+        term_put_tuple_element(frame_info, 1, term_from_int(current_offset));
+    } else {
+        // This branch allows overwriting found module name, with the name provided (module_atom)
+        // same applies to function_atom and arity. This allows having NIFs and BIFs on the
+        // stacktrace
+
+        term args_list = term_nil();
+        for (int arg_i = arity - 1; arg_i >= 0; arg_i--) {
+            args_list = term_list_prepend(ctx->x[arg_i], args_list, &ctx->heap);
+        }
+
+        frame_info = term_alloc_tuple(5, &ctx->heap);
+        term_put_tuple_element(frame_info, 0, term_from_int(mod->module_index));
+        term_put_tuple_element(frame_info, 1, term_from_int(current_offset));
+        term_put_tuple_element(frame_info, 2, module_atom);
+        term_put_tuple_element(frame_info, 3, function_atom);
+        term_put_tuple_element(frame_info, 4, args_list);
+    }
     raw_stacktrace = term_list_prepend(frame_info, raw_stacktrace, &ctx->heap);
 
     prev_mod = NULL;
@@ -264,6 +318,18 @@ static term find_path_created(const void *key, struct ModulePathPair *module_pat
     return term_invalid_term();
 }
 
+static bool is_bif_or_nif(
+    term module_atom, term function_atom, avm_int_t arity, const GlobalContext *glb)
+{
+    atom_index_t module_atom_index = term_to_atom_index(module_atom);
+    atom_index_t function_atom_index = term_to_atom_index(function_atom);
+    char mfa[MAX_MFA_NAME_LEN];
+    atom_table_write_mfa(
+        glb->atom_table, mfa, sizeof(mfa), module_atom_index, function_atom_index, arity);
+
+    return (bif_registry_get_handler(mfa) != NULL) || (nifs_get(mfa) != NULL);
+}
+
 term stacktrace_build(Context *ctx, term *stack_info, uint32_t live)
 {
     GlobalContext *glb = ctx->global;
@@ -308,6 +374,10 @@ term stacktrace_build(Context *ctx, term *stack_info, uint32_t live)
     int module_path_idx = 0;
     while (!term_is_nil(el)) {
         term mod_index_tuple = term_get_list_head(el);
+
+        size_t mod_index_tuple_arity = term_get_tuple_arity(mod_index_tuple);
+        assert((mod_index_tuple_arity == 2) || (mod_index_tuple_arity == 5));
+
         term cp = module_address(
             term_to_int(term_get_tuple_element(mod_index_tuple, 0)),
             term_to_int(term_get_tuple_element(mod_index_tuple, 1)));
@@ -318,9 +388,6 @@ term stacktrace_build(Context *ctx, term *stack_info, uint32_t live)
         module_cp_to_label_offset(cp, &cp_mod, &label, NULL, &mod_offset, ctx->global);
 
         term module_name = module_get_name(cp_mod);
-
-        term frame_i = term_alloc_tuple(4, &ctx->heap);
-        term_put_tuple_element(frame_i, 0, module_name);
 
         term aux_data = term_nil();
         if (module_has_line_chunk(cp_mod)) {
@@ -362,18 +429,46 @@ term stacktrace_build(Context *ctx, term *stack_info, uint32_t live)
                 aux_data = term_list_prepend(file_tuple, aux_data, &ctx->heap);
             }
         }
-        term_put_tuple_element(frame_i, 3, aux_data);
 
-        atom_index_t function_name;
-        int arity = 0;
-        bool result = module_get_function_from_label(cp_mod, label, &function_name, &arity);
+        term frame_i = term_alloc_tuple(4, &ctx->heap);
+        if (mod_index_tuple_arity == 5) {
+            term raw_module = term_get_tuple_element(mod_index_tuple, 2);
+            term raw_function = term_get_tuple_element(mod_index_tuple, 3);
+            term raw_arity_or_args = term_get_tuple_element(mod_index_tuple, 4);
+            term_put_tuple_element(frame_i, 0, raw_module);
+            term_put_tuple_element(frame_i, 1, raw_function);
+            term_put_tuple_element(frame_i, 2, raw_arity_or_args);
 
-        if (LIKELY(result)) {
-            term_put_tuple_element(frame_i, 1, term_from_atom_index(function_name));
-            term_put_tuple_element(frame_i, 2, term_from_int(arity));
+            avm_int_t arity;
+            if (term_is_int(raw_arity_or_args)) {
+                arity = term_to_int(raw_arity_or_args);
+            } else if (term_is_list(raw_arity_or_args)) {
+                int is_proper;
+                arity = term_list_length(raw_arity_or_args, &is_proper);
+                assert(is_proper);
+            } else {
+                AVM_ABORT();
+            }
+
+            if ((module_name == raw_module) && !is_bif_or_nif(raw_module, raw_function, arity, glb)) {
+                term_put_tuple_element(frame_i, 3, aux_data);
+            } else {
+                // mismatch means that error likely happened somewhere else, like a NIF, hence:
+                // aux_data is misleading and should be discarded
+                term_put_tuple_element(frame_i, 3, term_nil());
+            }
         } else {
-            term_put_tuple_element(frame_i, 1, UNDEFINED_ATOM);
-            term_put_tuple_element(frame_i, 2, term_from_int(0));
+            term_put_tuple_element(frame_i, 0, module_name);
+            atom_index_t function_name;
+            int arity = 0;
+            if (LIKELY(module_get_function_from_label(cp_mod, label, &function_name, &arity))) {
+                term_put_tuple_element(frame_i, 1, term_from_atom_index(function_name));
+                term_put_tuple_element(frame_i, 2, term_from_int(arity));
+            } else {
+                term_put_tuple_element(frame_i, 1, UNDEFINED_ATOM);
+                term_put_tuple_element(frame_i, 2, term_from_int(0));
+            }
+            term_put_tuple_element(frame_i, 3, aux_data);
         }
         stacktrace = term_list_prepend(frame_i, stacktrace, &ctx->heap);